+++ backend/src/main/java/ca/metricalsky/winston/client/CommentsDisabledException.java
@@ -1,6 +1,7 @@
 package ca.metricalsky.winston.client;
 
 import com.google.api.client.googleapis.json.GoogleJsonResponseException;
+import com.google.common.annotations.VisibleForTesting;
 import org.springframework.http.HttpStatus;
 
 import java.net.URI;
@@ -10,7 +11,7 @@ public class CommentsDisabledException extends YouTubeException {
     private static final URI TYPE = URI.create("/api/problem/comments-disabled");
     private static final String MESSAGE = "Comments are disabled for the requested video.";
 
-    CommentsDisabledException(GoogleJsonResponseException cause) {
+    public CommentsDisabledException(GoogleJsonResponseException cause) {
         super(HttpStatus.UNPROCESSABLE_ENTITY, MESSAGE, cause);
         getBody().setType(TYPE);
     }