+++ backend/src/main/java/ca/metricalsky/winston/service/fetch/FetchActionHandler.java
@@ -1,8 +1,0 @@
-package ca.metricalsky.winston.service.fetch;
-
-import ca.metricalsky.winston.entity.fetch.FetchAction;
-
-public interface FetchActionHandler<T> {
-
-    FetchResult<T> fetch(FetchAction fetchAction);
-}