+++ backend/src/test/java/ca/metricalsky/winston/service/fetch/AsyncFetchServiceTest.java
@@ -1,22 +1,26 @@
 package ca.metricalsky.winston.service.fetch;
 
 import ca.metricalsky.winston.dto.fetch.FetchRequestDto;
-import ca.metricalsky.winston.entity.fetch.FetchAction;
 import ca.metricalsky.winston.entity.fetch.FetchRequest;
+import ca.metricalsky.winston.entity.fetch.FetchRequest.FetchType;
 import ca.metricalsky.winston.events.FetchEvent;
 import ca.metricalsky.winston.events.PublisherException;
-import ca.metricalsky.winston.mapper.entity.FetchRequestMapper;
 import ca.metricalsky.winston.events.SsePublisher;
+import ca.metricalsky.winston.exception.AppException;
+import ca.metricalsky.winston.mapper.entity.FetchRequestMapper;
+import ca.metricalsky.winston.service.fetch.request.FetchRequestHandler;
+import ca.metricalsky.winston.service.fetch.request.FetchRequestHandlerFactory;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.http.HttpStatus;
 
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -26,124 +30,68 @@ class AsyncFetchServiceTest {
     private AsyncFetchService asyncFetchService;
 
     @Mock
-    private FetchActionHandler<Object> fetchActionHandler;
+    private FetchRequestHandler fetchRequestHandler;
     @Mock
-    private FetchActionHandlerFactory fetchActionHandlerFactory;
-    @Mock
-    private FetchActionService fetchActionService;
+    private FetchRequestHandlerFactory fetchRequestHandlerFactory;
     @Mock
     private FetchRequestMapper fetchRequestMapper;
-    @Mock
-    private FetchRequestService fetchRequestService;
     @Mock
     private SsePublisher ssePublisher;
 
     @Test
     void fetch() {
         var fetchRequestDto = new FetchRequestDto();
-        var fetchRequest = FetchRequest.builder()
-                .id(1L)
-                .fetchType(FetchRequest.FetchType.CHANNELS)
-                .objectId("objectId")
-                .build();
-        var fetchAction = FetchAction.builder()
-                .fetchRequestId(1L)
-                .actionType(FetchAction.ActionType.CHANNELS)
-                .objectId("objectId")
-                .status(FetchAction.Status.FETCHING)
-                .build();
-        var fetchResult = new FetchResult<>(fetchAction, new Object(), null);
+        var fetchRequest = new FetchRequest();
 
         when(fetchRequestMapper.toFetchRequest(fetchRequestDto))
                 .thenReturn(fetchRequest);
-        when(fetchRequestService.startFetch(fetchRequest))
-                .thenReturn(fetchRequest);
-        when(fetchActionService.actionFetching(any(FetchAction.class)))
-                .thenReturn(fetchAction);
-        when(fetchActionHandlerFactory.getHandlerForAction(fetchAction))
-                .thenReturn((FetchActionHandler) fetchActionHandler);
-        when(fetchActionHandler.fetch(fetchAction))
-                .thenReturn(fetchResult);
+        when(fetchRequestHandlerFactory.getHandler(fetchRequest))
+                .thenReturn(fetchRequestHandler);
 
         asyncFetchService.fetch(fetchRequestDto, ssePublisher);
 
-        verify(fetchActionService).actionCompleted(fetchAction, 1);
-        verify(ssePublisher).publish(any(FetchEvent.class));
-        verify(fetchRequestService).fetchCompleted(fetchRequest);
+        verify(fetchRequestHandler).fetch(fetchRequest, ssePublisher);
         verify(ssePublisher).complete();
     }
 
     @Test
-    void fetch_publisherException() {
+    void fetch_actionFailed() {
         var fetchRequestDto = new FetchRequestDto();
-        var fetchRequest = FetchRequest.builder()
-                .id(1L)
-                .fetchType(FetchRequest.FetchType.CHANNELS)
-                .objectId("objectId")
-                .build();
-        var fetchAction = FetchAction.builder()
-                .fetchRequestId(1L)
-                .actionType(FetchAction.ActionType.CHANNELS)
-                .objectId("objectId")
-                .status(FetchAction.Status.FETCHING)
-                .build();
-        var fetchResult = new FetchResult<>(fetchAction, new Object(), null);
+        var fetchRequest = FetchRequest.builder().fetchType(FetchType.CHANNELS).build();
+        var exception = new AppException(HttpStatus.INTERNAL_SERVER_ERROR, "");
 
         when(fetchRequestMapper.toFetchRequest(fetchRequestDto))
                 .thenReturn(fetchRequest);
-        when(fetchRequestService.startFetch(fetchRequest))
-                .thenReturn(fetchRequest);
-        when(fetchActionService.actionFetching(any(FetchAction.class)))
-                .thenReturn(fetchAction);
-        when(fetchActionHandlerFactory.getHandlerForAction(fetchAction))
-                .thenReturn((FetchActionHandler) fetchActionHandler);
-        when(fetchActionHandler.fetch(fetchAction))
-                .thenReturn(fetchResult);
-        doThrow(PublisherException.class)
-                .when(ssePublisher).publish(any(FetchEvent.class));
+        when(fetchRequestHandlerFactory.getHandler(fetchRequest))
+                .thenReturn(fetchRequestHandler);
+        doThrow(exception)
+                .when(fetchRequestHandler).fetch(fetchRequest, ssePublisher);
+        when(ssePublisher.isOpen())
+                .thenReturn(true);
 
         asyncFetchService.fetch(fetchRequestDto, ssePublisher);
 
-        verify(fetchActionService).actionCompleted(fetchAction, 1);
         verify(ssePublisher).publish(any(FetchEvent.class));
-        verify(fetchRequestService).fetchCompleted(fetchRequest);
+        verify(ssePublisher).completeWithError(exception);
     }
 
     @Test
-    void fetch_publisherException_hasNextFetchAction() {
+    void fetch_publisherClosed() {
         var fetchRequestDto = new FetchRequestDto();
-        var fetchRequest = FetchRequest.builder()
-                .id(1L)
-                .fetchType(FetchRequest.FetchType.CHANNELS)
-                .objectId("objectId")
-                .build();
-        var fetchAction = FetchAction.builder()
-                .fetchRequestId(1L)
-                .actionType(FetchAction.ActionType.CHANNELS)
-                .objectId("objectId")
-                .status(FetchAction.Status.FETCHING)
-                .build();
-        var nextFetchAction = new FetchAction();
-        var fetchResult = new FetchResult<>(fetchAction, new Object(), nextFetchAction);
+        var fetchRequest = FetchRequest.builder().fetchType(FetchType.CHANNELS).build();
+        var exception = new PublisherException("");
 
         when(fetchRequestMapper.toFetchRequest(fetchRequestDto))
                 .thenReturn(fetchRequest);
-        when(fetchRequestService.startFetch(fetchRequest))
-                .thenReturn(fetchRequest);
-        when(fetchActionService.actionFetching(any(FetchAction.class)))
-                .thenReturn(fetchAction);
-        when(fetchActionHandlerFactory.getHandlerForAction(fetchAction))
-                .thenReturn((FetchActionHandler) fetchActionHandler);
-        when(fetchActionHandler.fetch(fetchAction))
-                .thenReturn(fetchResult);
-        doThrow(PublisherException.class)
-                .when(ssePublisher).publish(any(FetchEvent.class));
+        when(fetchRequestHandlerFactory.getHandler(fetchRequest))
+                .thenReturn(fetchRequestHandler);
+        doThrow(exception)
+                .when(fetchRequestHandler).fetch(fetchRequest, ssePublisher);
+        when(ssePublisher.isOpen())
+                .thenReturn(false);
 
         asyncFetchService.fetch(fetchRequestDto, ssePublisher);
 
-        verify(fetchActionService).actionCompleted(fetchAction, 1);
-        verify(ssePublisher).publish(any(FetchEvent.class));
-        verify(fetchActionService).actionReady(nextFetchAction);
-        verify(fetchRequestService).fetchFailed(eq(fetchRequest), any(PublisherException.class));
+        verifyNoMoreInteractions(ssePublisher);
     }
 }