+++ backend/src/test/java/ca/metricalsky/winston/service/fetch/action/FetchCommentsActionHandlerTest.java
@@ -1,28 +1,36 @@
-package ca.metricalsky.winston.service.fetch;
+package ca.metricalsky.winston.service.fetch.action;
 
 import ca.metricalsky.winston.client.CommentsDisabledException;
 import ca.metricalsky.winston.client.YouTubeClientAdapter;
 import ca.metricalsky.winston.entity.Video;
 import ca.metricalsky.winston.entity.fetch.FetchAction;
+import ca.metricalsky.winston.events.FetchEvent;
+import ca.metricalsky.winston.events.FetchStatus;
+import ca.metricalsky.winston.events.SsePublisher;
 import ca.metricalsky.winston.repository.VideoRepository;
 import ca.metricalsky.winston.service.CommentService;
+import ca.metricalsky.winston.service.fetch.FetchActionService;
 import com.google.api.services.youtube.model.Comment;
 import com.google.api.services.youtube.model.CommentThread;
 import com.google.api.services.youtube.model.CommentThreadListResponse;
 import com.google.api.services.youtube.model.CommentThreadSnippet;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.http.HttpStatus;
 
 import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.catchThrowableOfType;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -35,11 +43,17 @@ class FetchCommentsActionHandlerTest {
     private FetchCommentsActionHandler fetchCommentsActionHandler;
 
     @Mock
+    private FetchActionService fetchActionService;
+    @Mock
     private CommentService commentService;
     @Mock
     private VideoRepository videoRepository;
     @Mock
     private YouTubeClientAdapter youTubeClientAdapter;
+    @Mock
+    private SsePublisher ssePublisher;
+    @Captor
+    private ArgumentCaptor<FetchEvent> fetchEvent;
 
     @Test
     void fetch() {
@@ -47,21 +61,29 @@ class FetchCommentsActionHandlerTest {
                 .actionType(FetchAction.ActionType.COMMENTS)
                 .objectId(VIDEO_ID)
                 .build();
+        var commentThreadListResponse = buildCommentThreadListResponse();
+
+        when(fetchActionService.actionFetching(fetchAction))
+                .thenReturn(fetchAction);
         when(youTubeClientAdapter.getComments(fetchAction))
-                .thenReturn(buildCommentThreadListResponse());
+                .thenReturn(commentThreadListResponse);
 
-        var fetchResult = fetchCommentsActionHandler.fetch(fetchAction);
+        var nextFetchAction = fetchCommentsActionHandler.fetch(fetchAction, ssePublisher);
 
-        assertThat(fetchResult)
-                .hasFieldOrPropertyWithValue("actionType", FetchAction.ActionType.COMMENTS)
-                .hasFieldOrPropertyWithValue("objectId", VIDEO_ID)
-                .hasFieldOrPropertyWithValue("nextFetchAction", null);
-        assertThat(fetchResult.items())
-                .hasSize(1)
-                .first()
-                .hasFieldOrPropertyWithValue("id", COMMENT_ID);
+        assertThat(nextFetchAction).isNull();
 
         verify(commentService).saveAll(anyList());
+        verify(fetchActionService).actionCompleted(fetchAction, commentThreadListResponse.getItems().size());
+        verify(ssePublisher).publish(fetchEvent.capture());
+
+        assertThat(fetchEvent.getValue())
+                .hasFieldOrPropertyWithValue("type", "fetch-comments")
+                .hasFieldOrPropertyWithValue("objectId", VIDEO_ID)
+                .hasFieldOrPropertyWithValue("status", FetchStatus.COMPLETED);
+        assertThat(fetchEvent.getValue().items())
+                .hasSize(commentThreadListResponse.getItems().size())
+                .first()
+                .hasFieldOrPropertyWithValue("id", COMMENT_ID);
     }
 
     @Test
@@ -73,14 +95,24 @@ class FetchCommentsActionHandlerTest {
         var video = new Video();
         video.setId(VIDEO_ID);
 
-        when(youTubeClientAdapter.getComments(fetchAction)).thenThrow(CommentsDisabledException.class);
-        when(videoRepository.findById(VIDEO_ID)).thenReturn(Optional.of(video));
+        when(fetchActionService.actionFetching(fetchAction))
+                .thenReturn(fetchAction);
+        when(youTubeClientAdapter.getComments(fetchAction))
+                .thenThrow(new CommentsDisabledException(null));
+        when(videoRepository.findById(VIDEO_ID))
+                .thenReturn(Optional.of(video));
 
-        assertThatThrownBy(() -> fetchCommentsActionHandler.fetch(fetchAction))
-                .isExactlyInstanceOf(CommentsDisabledException.class);
+        var exception = catchThrowableOfType(CommentsDisabledException.class,
+                () -> fetchCommentsActionHandler.fetch(fetchAction, ssePublisher));
 
-        assertThat(video.isCommentsDisabled()).isTrue();
+        assertThat(exception)
+                .hasFieldOrPropertyWithValue("status", HttpStatus.UNPROCESSABLE_ENTITY)
+                .hasMessageEndingWith("Comments are disabled for the requested video.");
+        assertThat(video.isCommentsDisabled())
+                .isTrue();
+
         verify(videoRepository).save(video);
+        verifyNoInteractions(ssePublisher);
     }
 
     private static CommentThreadListResponse buildCommentThreadListResponse() {